# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is API of movie booking service",
    "version": "0.1.0",
    "title": "Movie booking API"
  },
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "manager"
        ],
        "summary": "Get all users",
        "description": "",
        "operationId": "getUsers",
        "consumes": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create new user",
        "description": "",
        "operationId": "createUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "405": {
            "description": "Validation exception"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/users/list": {
      "get": {
        "tags": [
          "manager"
        ],
        "summary": "Finds users by pagination",
        "description": "",
        "operationId": "getUsersByPagination",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page number",
            "required": false,
            "type": "string",
            "default": "1"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "limit of users",
            "required": false,
            "type": "string",
            "default": "3"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "order list of users",
            "required": false,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Cinema"
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "manager"
        ],
        "summary": "Find user by ID",
        "description": "Returns a single user",
        "operationId": "getUserById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update user",
        "description": "",
        "operationId": "updateUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to update",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "User object that needs to be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "405": {
            "description": "Validation exception"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "manager"
        ],
        "summary": "Deletes a user",
        "description": "",
        "operationId": "removeUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/invite": {
      "get": {
        "tags": [
          "manager"
        ],
        "summary": "Invite new manager",
        "description": "",
        "operationId": "inviteManager",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "invite user email",
            "required": true,
            "type": "string",
            "default": "user@gmail.com"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "405": {
            "description": "Validation exception"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/manager": {
      "post": {
        "tags": [
          "manager"
        ],
        "summary": "Create new manager",
        "description": "",
        "operationId": "createManager",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "invite code",
            "required": true,
            "type": "number"
          },
          {
            "in": "body",
            "name": "body",
            "description": "create new manager",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "405": {
            "description": "Validation exception"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "Cinema": {
      "type": "object",
      "required": [
        "title"
      ],
      "properties": {
        "title": {
          "type": "string",
          "example": "New cinema"
        },
        "address": {
          "type": "string",
          "example": "Simple address"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Smith"
        },
        "email": {
          "type": "string",
          "example": "user@gmail.com"
        },
        "password": {
          "type": "password",
          "example": "user"
        },
        "phoneNumber": {
          "type": "string",
          "example": "+38012345"
        }
      }
    },
...
  }
}